BUG: Inconsistent DLL linkage in MeshContourSegmentation
When the module has been renamed (r19175), the EXPORT variable hasn't been renamed accordingly.

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=19529
git-svn-id: http://svn.slicer.org/Slicer4/trunk@19529 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/CLI/MeshContourSegmentation/ModuleEntry.cxx
@@ -80,7 +80,7 @@ void MeshContourEvolver::entry_main( vtkPolyData* inputMesh,
   double                       thispt[3];
   vtkSmartPointer<vtkIntArray> initialPoints = vtkSmartPointer<vtkIntArray>::New();
   unsigned int                 iNumMeshVerts  = verts->GetNumberOfPoints();
-  unsigned int                 iNumSeedPoints = initPoints3D.size();
+  size_t                       iNumSeedPoints = initPoints3D.size();
   std::cout << "appending " << iNumSeedPoints << " points to list. \n";
 
   if( iNumSeedPoints < 2 )
